Always return a list from Genderize.get()
This fixes the actual problem behind PR #2, which is that the API
returns a single object when a single name is requested, and a list
otherwise. The single object is now promoted to a list.

The get1() method is tested now, which covers the single-object case.

This change also makes `test.py` nonexecutable. It was previously
executable, so `nosetests` was ignoring it, and thus testing nothing:
https://nose.readthedocs.org/en/latest/usage.html#extended-usage